+++ include/source_output_effects_ui.hpp
@@ -13,6 +13,7 @@
 #include "pitch_ui.hpp"
 #include "reverb_ui.hpp"
 #include "source_output_effects.hpp"
+#include "stereo_tools_ui.hpp"
 #include "webrtc_ui.hpp"
 
 class SourceOutputEffectsUi : public Gtk::Box, public EffectsBaseUi {
@@ -48,6 +49,7 @@ class SourceOutputEffectsUi : public Gtk::Box, public EffectsBaseUi {
   WebrtcUi* webrtc_ui = nullptr;
   MultibandCompressorUi* multiband_compressor_ui = nullptr;
   MultibandGateUi* multiband_gate_ui = nullptr;
+  StereoToolsUi* stereo_tools_ui = nullptr;
 
   void level_meters_connections();
   void up_down_connections();